+++ Options/OptionsManager.cs
@@ -34,9 +34,15 @@ public sealed class OptionsManager
         }
 
         new ConfigOption<bool>(
-            key: nameof(OptionType.AutoCleanupEnabled),
-            displayName: "Enable Automatic Cleanup",
-            defaultValue: defaults.autoCleanupEnabled,
+            key: nameof(OptionType.AutoCleanupCableSpoolsEnabled),
+            displayName: "Enable Automatic Cleanup of empty Cable Spools",
+            defaultValue: defaults.autoCleanupCableSpoolsEnabled,
+            description: "Enables automatic cleanup of empty cable spools.");
+
+        new ConfigOption<bool>(
+            key: nameof(OptionType.AutoCleanupSFPBoxesEnabled),
+            displayName: "Enable Automatic Cleanup of empty SFP Boxes",
+            defaultValue: defaults.autoCleanupSFPBoxesEnabled,
             description: "Enables automatic cleanup of empty SFP boxes.");
 
         new ConfigOption<int>(
@@ -44,6 +50,12 @@ public sealed class OptionsManager
             displayName: "Automatic Cleanup Interval",
             defaultValue: Math.Max(1, (int)Math.Round(defaults.autoCleanupIntervalMinutes)),
             description: "How often empty SFP boxes will be cleaned up, in minutes.");
+
+        new ConfigOption<float>(
+            key: nameof(OptionType.CableSpoolLengthThreshold),
+            displayName: "Cable Spool Length Threshold",
+            defaultValue: Math.Max(0.1f, defaults.cableSpoolLengthThreshold),
+            description: "The minimum length of cable spools to keep, in meters.");
 
         Initialized = true;
     }
@@ -59,6 +71,7 @@ public sealed class OptionsManager
         {
             var type when type == typeof(bool) => (T)(object)ModConfigSystem.GetBoolValue(TrashCleanerMod.ModName, optionType.ToString()),
             var type when type == typeof(int) => (T)(object)ModConfigSystem.GetIntValue(TrashCleanerMod.ModName, optionType.ToString()),
+            var type when type == typeof(float) => (T)(object)ModConfigSystem.GetFloatValue(TrashCleanerMod.ModName, optionType.ToString()),
             _ => default
         };
     }